AccelerationStructureManager, DescriptorManager, RayTracingPipeline
