--- src/renderer/AccelerationStructureUtils.cpp
@@ -1,101 +1,0 @@
-#include "pch.h"
-#include "AccelerationStructureUtils.hpp"
-
-void VRTR::AS::primitiveToGeometry(const std::vector<VertexRT> &vertices,
-                                   const std::vector<uint32_t> &indices,
-                                   std::unique_ptr<Buffer> &vertexBuffer,
-                                   std::unique_ptr<Buffer> &indexBuffer,
-                                   vk::AccelerationStructureGeometryKHR &geometry,
-                                   vk::AccelerationStructureBuildRangeInfoKHR &offsetInfo,
-                                   vk::Format vertexFormat,
-                                   vk::IndexType indexType)
-{
-    uint32_t triangleCount = static_cast<uint32_t>(indices.size() / 3U);
-
-    vk::AccelerationStructureGeometryTrianglesDataKHR triangles{
-        .pNext = nullptr,
-        .vertexFormat = vertexFormat,
-        .vertexData = vk::DeviceOrHostAddressConstKHR{vertexBuffer->getDeviceAddress()},
-        .vertexStride = sizeof(VertexRT),
-        .maxVertex = static_cast<uint32_t>(vertices.size() - 1),
-        .indexType = indexType,
-        .indexData = vk::DeviceOrHostAddressConstKHR{indexBuffer->getDeviceAddress()},
-        .transformData = {}};
-
-    geometry = vk::AccelerationStructureGeometryKHR{
-        .pNext = nullptr,
-        .geometryType = vk::GeometryTypeKHR::eTriangles,
-        .geometry = triangles,
-        .flags = vk::GeometryFlagBitsKHR::eOpaque | vk::GeometryFlagBitsKHR::eNoDuplicateAnyHitInvocation};
-
-    offsetInfo = vk::AccelerationStructureBuildRangeInfoKHR{
-        .primitiveCount = triangleCount,
-        .primitiveOffset = 0,
-        .firstVertex = 0,
-        .transformOffset = 0};
-}
-
-void VRTR::AS::createAccelerationStructure(VRTR::VULKAN_CONTEXT &ctx,
-                                           vk::AccelerationStructureTypeKHR asType,
-                                           VRTR::AS::AccelerationStructure &as,
-                                           vk::AccelerationStructureGeometryKHR &asGeometry,
-                                           vk::AccelerationStructureBuildRangeInfoKHR &asBuildRangeInfo,
-                                           vk::BuildAccelerationStructureFlagsKHR flags)
-{
-    // gemoetry data
-    vk::AccelerationStructureBuildGeometryInfoKHR asBuildInfo{
-        .type = asType,
-        .flags = flags,
-        .mode = vk::BuildAccelerationStructureModeKHR::eBuild,
-        .geometryCount = 1,
-        .pGeometries = &asGeometry,
-    };
-
-    std::vector<uint32_t> maxPrimCount(1);
-    maxPrimCount.at(0) = asBuildRangeInfo.primitiveCount;
-
-    // getting size info
-    // size needed for the acceleration structure itself,
-    // size needed for the scratch buffer
-    // size needed for the update scratch buffer
-    vk::AccelerationStructureBuildSizesInfoKHR asSizeInfo = ctx.logicalDevice.getAccelerationStructureBuildSizesKHR(
-        vk::AccelerationStructureBuildTypeKHR::eDevice,
-        asBuildInfo,
-        maxPrimCount);
-
-    vk::DeviceSize scratchSize = asSizeInfo.buildScratchSize;
-    vk::DeviceSize minAsScratchOffsetAlignment = ctx.asProperties.minAccelerationStructureScratchOffsetAlignment;
-    scratchSize = utils::aligned_size(scratchSize, minAsScratchOffsetAlignment);
-
-    Buffer scratchBuffer(ctx, BufferType::SCRATCH, scratchSize);
-
-    // we need also a buffer that will hold the acceleration structure
-    as.asBuffer = std::make_unique<Buffer>(ctx.logicalDevice, ctx.gpu, asSizeInfo.accelerationStructureSize,
-                                         vk::BufferUsageFlags{}, vk::MemoryPropertyFlagBits::eDeviceLocal,
-                                         vk::BufferUsageFlagBits2::eAccelerationStructureStorageKHR |
-                                             vk::BufferUsageFlagBits2::eShaderDeviceAddress);
-
-    vk::AccelerationStructureCreateInfoKHR asCreateInfo{
-        .pNext = nullptr,
-        .createFlags = {},
-        .buffer = as.asBuffer->getBuffer(),
-        .offset = 0,
-        .size = asSizeInfo.accelerationStructureSize,
-        .type = asType,
-        .deviceAddress = 0};
-    as.handle = vk::raii::AccelerationStructureKHR(ctx.logicalDevice, asCreateInfo);
-
-    asBuildInfo.dstAccelerationStructure = *as.handle;
-    asBuildInfo.scratchData = scratchBuffer.getDeviceAddress();
-
-    std::unique_ptr<TempCMDBufferManager> tempCmdBufferManager = std::make_unique<TempCMDBufferManager>(ctx.logicalDevice, ctx.queue, ctx.graphics_queue_index);
-    vk::raii::CommandBuffer& tempCmdBuffer = tempCmdBufferManager->createTempCmdBuffer();
-
-    std::array<vk::AccelerationStructureBuildRangeInfoKHR *, 1> BuildRangeInfos = {&asBuildRangeInfo};
-    tempCmdBuffer.buildAccelerationStructuresKHR({asBuildInfo}, BuildRangeInfos);
-
-    as.deviceAddress = ctx.logicalDevice.getAccelerationStructureAddressKHR(
-        vk::AccelerationStructureDeviceAddressInfoKHR{
-            .accelerationStructure = *as.handle});
-    tempCmdBufferManager->submitAndWaitTempCmdBuffer();
-}
