+++ src/renderer/buffer.cpp
@@ -56,6 +56,7 @@ namespace VRTR
 
         bufferMemory = vk::raii::DeviceMemory{logicalDevice, memoryAllocateInfo};
         buffer.bindMemory(bufferMemory, 0);
+        // tu nie musi byc getBufferAddressKHR?
         deviceAddress = logicalDevice.getBufferAddress(vk::BufferDeviceAddressInfo{.buffer=buffer});
     }
