+++ src/renderer/include/Camera.hpp
@@ -11,6 +11,7 @@ namespace VRTR
 
         struct Matrices
         {
+            // glm::mat4 model;
             glm::mat4 view;
             glm::mat4 perspective;
         } matrices;
